Transform FB feed to ActivityPub activity JSON
Upgrade Ruby verion to be able to use Hash#dig

Refactor Salvager class

Needed to update the loop iteration to not write each batch as its own array to
the output file. Also refactored #collect_album method to utlize the same loop
logic. Not a massive fan of the #send used now but it seemed the best way to
DRY up the code now that each file containing an array needs explicit handling
of the square brackets.

Add activity feed transformation

Comments in the code explain some of the mapping decisions. I've also modified
links and IDs.